software-design/02-api/requirements.txt
pydantic
plotly
pandas
vk_api
PyYAML
loguru